% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{binary2stars}
\alias{binary2stars}
\title{Converts response of endpoint /udf/binary to stars}
\usage{
binary2stars(json_out, file = FALSE)
}
\arguments{
\item{json_out}{String representing JSON or path to JSON}

\item{file}{Logical. Needs to be TRUE if \code{json_out} represents the path to a JSON file}
}
\value{
A \code{stars} object created out of the JSON
}
\description{
Converts response of endpoint /udf/binary to stars
}
