% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTful_services.R
\name{run_UDF.binary}
\alias{run_UDF.binary}
\title{UDFs on a \code{stars} object created from a binary string}
\usage{
run_UDF.binary(req)
}
\arguments{
\item{req}{The incoming HTTP POST request}
}
\value{
The response to the HTTP POST request
}
\description{
Run user-defined functions on a \code{stars} object created from a base64 encoded
string representing a ZIP file containing generic GeoTIFF files. The \code{stars}
is exposed as it is to the UDF.

Creates \code{stars} object from legend file and GeoTIFFs
}
