% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTful_services.R
\name{run_UDF.json.raw}
\alias{run_UDF.json.raw}
\title{Run user-defined functions on a \code{stars} object created from JSON arrays
exposed as it is to the UDF.}
\usage{
run_UDF.json.raw(req)
}
\arguments{
\item{req}{The incoming HTTP POST request}
}
\value{
The response to the HTTP POST request
}
\description{
This function is linked to the endpoint \code{/udf/raw}
}
