% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-based_services.R
\name{read_generics}
\alias{read_generics}
\title{Read legend file and dimensionality of incoming data}
\usage{
read_generics(legend_file, dimensionality)
}
\arguments{
\item{legend_file}{CSV file containing data on the incoming images which acts as a look-up table}

\item{dimensionality}{Vector of dimensions {x, y, band, time, raster/vector?}}
}
\value{
\code{stars} object returned by \link{bin_read_legend}
}
\description{
Read legend file and dimensionality of incoming data
}
