% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTful_services.R
\name{run_UDF.json}
\alias{run_UDF.json}
\title{UDF on \code{stars} object exposed as a list}
\usage{
run_UDF.json(req)
}
\arguments{
\item{req}{The incoming HTTP POST request}
}
\value{
The response to the HTTP POST request
}
\description{
Runs user-defined functions on a \code{stars} object created from JSON arrays
exposed as a list to the UDF.

This function is linked to the endpoint \code{/udf}
}
